requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
pika==1.2.0